+++ src/Pyro4/socketutil.py
@@ -232,9 +232,9 @@ def sendData(sock, data):
 _GLOBAL_DEFAULT_TIMEOUT = object()
 
 
-def createSocket(bind=None, connect=None, reuseaddr=False, keepalive=True, timeout=_GLOBAL_DEFAULT_TIMEOUT, noinherit=False, ipv6=False, nodelay=False):
+def createSocket(bind=None, connect=None, reuseaddr=False, keepalive=True, timeout=_GLOBAL_DEFAULT_TIMEOUT, noinherit=False, ipv6=False, nodelay=True):
     """
-    Create a socket. Default socket options are keepalive and IPv4 family.
+    Create a socket. Default socket options are keepalive and IPv4 family, and nodelay (nagle disabled).
     If 'bind' or 'connect' is a string, it is assumed a Unix domain socket is requested.
     Otherwise, a normal tcp/ip socket is used.
     Set ipv6=True to create an IPv6 socket rather than IPv4.